Add max trials.

+++ lib/lock.dart
@@ -8,6 +8,29 @@ import 'dart:io';
 import 'dart:async';
 import 'package:path/path.dart' as p;
 
+class LockException implements Exception {
+
+    /** Description of the cause of the timeout. */
+    final String message;
+    /** The max count that was exceeded. */
+    final num count;
+
+    LockException(this.message, [this.count]);
+
+    String toString() {
+      if (message != null) {
+        if (count != null) return "LockException after $count trials: $message";
+        return "LockException: $message";
+      }
+      if (count != null) return "LockException after $count trials";
+      return "LockException";
+    }
+
+}
+
+
+Set _locks = new Set();
+
 get _baseDir => p.dirname(Platform.script.toFilePath());
 
 _lockPath(path) {
@@ -20,21 +43,32 @@ bool tryLock(path) {
     return false;
   } else {
     new File(_lockPath(path)).createSync();
+    _locks.add(path);
     return true;
   }
 }
 
 void freeLock(path) {
- new File(_lockPath(path)).deleteSync();
+  if (_locks.contains(path)) {
+    new File(_lockPath(path)).deleteSync();
+  }
 }
 
-Future obtainLock(path, {Duration tryInterval: const Duration(seconds: 1)}) {
+Future obtainLock(path, {Duration tryInterval: const Duration(seconds: 1),
+                         num maxTrials: -1}) {
   var completer = new Completer();
 
+  var trials = 0;
   tryCallback(Timer timer) {
     if (tryLock(path)) {
       timer.cancel();
       completer.complete();
+    } else {
+      trials++;
+      if (trials == maxTrials) {
+        timer.cancel();
+        completer.completeError(new LockException("Could not obtain lock", trials));
+      }
     }
   }
   var timer = new Timer.periodic(tryInterval, tryCallback);
@@ -44,9 +78,9 @@ Future obtainLock(path, {Duration tryInterval: const Duration(seconds: 1)}) {
   return completer.future;
 }
 
-Future _runWithLock(path, Future callback(), tryInterval) {
+Future _runWithLock(path, Future callback(), tryInterval, maxTrials) {
   var value;
-  return obtainLock(path, tryInterval: tryInterval).then((_) {
+  return obtainLock(path, tryInterval: tryInterval, maxTrials: maxTrials).then((_) {
     var listeners = [];
     var signals = [
                    // BUG in Dart -- can not listen to all three signals
@@ -70,8 +104,11 @@ Future _runWithLock(path, Future callback(), tryInterval) {
   });
 }
 
-Future runWithLock(path, Future callback(), {Duration tryInterval: const Duration(seconds: 1)}) {
-  return runZoned(() => _runWithLock(path, callback, tryInterval), onError: (e, s) {
+Future runWithLock(path, Future callback(),
+                   {Duration tryInterval: const Duration(seconds: 1),
+                    num maxTrials: -1}) {
+  return runZoned(() => _runWithLock(path, callback, tryInterval, maxTrials),
+                                     onError: (e, s) {
     freeLock(path);
     print(e);
     print(s);
